@@ -27,9 +27,8 @@
 
         public Task<TEntity> GetByIdWithDeletedAsync(params object[] id)
         {
-            var byIdPredicate = EfExpressionHelper.BuildByIdPredicate<TEntity>(this.Context, id);
-
-            return this.AllWithDeleted().FirstOrDefaultAsync(byIdPredicate);
+            var getByIdPredicate = EfExpressionHelper.BuildByIdPredicate<TEntity>(this.Context, id);
+            return this.AllWithDeleted().FirstOrDefaultAsync(getByIdPredicate);
         }
 
         public void HardDelete(TEntity entity) => base.Delete(entity);
@@ -38,7 +37,6 @@
         {
             entity.IsDeleted = false;
             entity.DeletedOn = null;
-
             this.Update(entity);
         }
 
@@ -46,7 +44,6 @@
         {
             entity.IsDeleted = true;
             entity.DeletedOn = DateTime.UtcNow;
-
             this.Update(entity);
         }
     }
