@@ -1,10 +1,11 @@
 ﻿namespace AspNetCoreTemplate.Services.Data
 {
     using System.Collections.Generic;
+    using System.Threading.Tasks;
 
     public interface ISettingsService
     {
-        int GetCount();
+        Task<int> GetCountAsync();
 
         IEnumerable<T> GetAll<T>();
     }
