@@ -1,5 +1,9 @@
 ﻿namespace AspNetCoreTemplate.Web.Controllers
 {
+    using System.Diagnostics;
+
+    using AspNetCoreTemplate.Web.ViewModels;
+
     using Microsoft.AspNetCore.Mvc;
 
     public class HomeController : BaseController
@@ -15,6 +19,10 @@
         }
 
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-        public IActionResult Error() => this.View();
+        public IActionResult Error()
+        {
+            return this.View(
+                new ErrorViewModel { RequestId = Activity.Current?.Id ?? this.HttpContext.TraceIdentifier });
+        }
     }
 }
